reference/9_AGENT_PROTOCOL/pyproject.toml
[project]
name = "google-adk-study"
version = "0.1.0"
description = "Hello Google ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.116.2",
    "fastmcp==2.13.0.1",
    "google-adk[a2a]==1.14.1",
    "google-genai==1.38.0",
    "litellm==1.77.1",
    "openai==1.108.1",
    "pillow==12.0.0",
    "python-dotenv==1.1.1",
    "uvicorn==0.36.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
